chore: rename sso/login to update matkul
